# packages/example-schemas/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Example REST API",
    "version": "0.1.0",
    "description": "Stubbed REST API demonstrating the schemas/service/client monorepo pattern."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Shared secret. Required on operations declaring ApiKeyAuth."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "info": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          }
        },
        "required": [
          "error",
          "message"
        ],
        "description": "Standard error response shape from `@polygonlabs/express`'s `createErrorHandler`. `info` is present when the underlying `HTTPError` carries structured info (e.g. validation failures, domain-specific context); absent for plain errors and most non-validation HTTPErrors."
      },
...
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "info": {
            "$ref": "#/components/schemas/ValidationErrorInfo"
          }
        },
        "required": [
          "error",
          "message",
          "info"
        ],
        "description": "Narrowed `ErrorResponse` for 400s emitted by the registry-driven router when request validation fails. `info` is non-optional and carries the section-keyed `ValidationErrorInfo` shape."
      },
      "ValidationErrorInfo": {
        "type": "object",
        "properties": {
          "params": {
            "$ref": "#/components/schemas/ZodErrorTree"
          },
          "query": {
            "$ref": "#/components/schemas/ZodErrorTree"
          },
          "body": {
            "$ref": "#/components/schemas/ZodErrorTree"
          },
          "headers": {
            "$ref": "#/components/schemas/ZodErrorTree"
          }
        },
        "description": "Section-keyed map of request validation failures. Only sections that failed appear; each value is the `z.treeifyError` tree for that section. Clients can wire e.g. `info.body.properties.label.errors[0]` directly into field-level UI feedback rather than walking a flat issue list."
      },
      "ZodErrorTree": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ZodErrorTree"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZodErrorTree"
            }
          }
        },
        "required": [
          "errors"
        ],
        "description": "Recursive Zod validation error tree (the output of `z.treeifyError`). `errors` carries the messages at this node; `properties` keys nested object errors by field name; `items` indexes nested array errors."
      },
...
      "EventList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexedEvent"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "IndexedEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chain": {
            "type": "integer"
          },
          "contractAddress": {
            "type": "string"
          },
          "eventName": {
            "type": "string"
          },
          "blockNumber": {
            "type": "integer"
          },
          "transactionIndex": {
            "type": "integer"
          },
          "txHash": {
            "type": "string"
          },
          "logIndex": {
            "type": "integer"
          },
          "args": {
            "type": "object",
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            }
          },
          "indexedAt": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "chain",
          "contractAddress",
          "eventName",
          "blockNumber",
          "txHash",
          "logIndex",
          "args",
          "indexedAt"
        ]
      }
    },
...
  },
  "paths": {
...
    "/events": {
      "get": {
        "operationId": "listEvents",
        "summary": "List indexed on-chain events",
        "description": "Returns decoded events indexed by example-indexer, newest-first, with optional chain / contractAddress / eventName filters and opaque-cursor pagination.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "nullable": true
            },
            "required": false,
            "name": "chain",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "contractAddress",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "eventName",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Page of indexed events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "400": {
            "description": "Request failed schema validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Malformed pagination cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}